requirements.txt
blinker==1.4
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.4
python-dotenv==0.17.1
six==1.16.0
Werkzeug==1.0.1
